fix minor bugs

+++ src/scrape/scraper.py
@@ -90,6 +90,8 @@ class DownloadProgressBar(tqdm):
 
 
 def download_url(url, output_path):
+    if os.path.exists(output_path):
+        return True
     with DownloadProgressBar(unit='B', unit_scale=True, miniters=1, desc=url.split('/')[-1]) as t:
         urllib.request.urlretrieve(url, filename=output_path, reporthook=t.update_to)
